Add support for returning pathlib.Path objects
With this commit the "as_path" keyword argument is being added. In order to
keep backward compatibility, its default value is "False"  If it is "True",
then the returned path is converted to an instance of "pathlib.Path". On
the functions that support "multipath", if "multipath" is enabled then
"as_path" is ignored.

On Python 2, if "pathlib2" is *not* installed, then "as_path" does nothing.